requirements.txt
ccxt==4.2.25
websocket-client==1.7.0
pandas==2.1.4
numpy==1.24.3
ta-lib==0.4.28
scikit-learn==1.3.2
tensorflow==2.15.0
requests==2.31.0
python-telegram-bot==20.7
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
python-dotenv==1.0.0
schedule==1.2.0
asyncio==3.4.3
aiohttp==3.9.1
websockets==12.0
plotly==5.17.0
dash==2.14.2
dash-bootstrap-components==1.5.0
cryptography==41.0.8
pycryptodome==3.19.0
deap==1.4.1
joblib==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
textblob==0.17.1